cont init

+++ src/IShape.java
@@ -6,9 +6,9 @@ public interface IShape {
 	}
 	
 	public void color(Color color);
-	public float get_area();
+	public double get_area();
 	public void delete();
-	public void move(float change_x, float change_y);
+	public void move(double change_x, double change_y);
 	public void copy();
 	public void circumference();
 	public void is_inside();
